scripts/upload_pfcc_ct_morningqa_catphan/catphan_image_analysis/fastapi_server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pymongo==4.6.0
redis==5.0.1
rq==1.15.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
SimpleITK==2.3.1
pydicom==2.4.4
numpy==1.24.3
aiofiles==23.2.1
